+++ packages/ui-shared/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - **Constructor**
     - create ButtonLinkBox component
+    - create TitleDescriptionBox component
 
 ## Release 1.0.30
 - **package.json**